move common type defs to a separate package
