+++ server/push/fcm/payload.go
@@ -3,16 +3,15 @@ package fcm
 import (
 	"encoding/json"
 	"errors"
-	"reflect"
 	"strconv"
 	"time"
 
-	legacy "firebase.google.com/go/messaging"
 	fcmv1 "google.golang.org/api/fcm/v1"
 
 	"github.com/tinode/chat/server/drafty"
 	"github.com/tinode/chat/server/logs"
 	"github.com/tinode/chat/server/push"
+	"github.com/tinode/chat/server/push/common"
 	"github.com/tinode/chat/server/store"
 	t "github.com/tinode/chat/server/store/types"
 )
@@ -23,229 +22,6 @@ const (
 	// TTL of a regular push notification in seconds.
 	defaultTimeToLive = 3600
 )
-
-// Payload to be sent for a specific notification type.
-type commonPayload struct {
-	// Common for APNS and Android
-	Body         string   `json:"body,omitempty"`
-	Title        string   `json:"title,omitempty"`
-	TitleLocKey  string   `json:"title_loc_key,omitempty"`
-	TitleLocArgs []string `json:"title_loc_args,omitempty"`
-
-	// Android
-	BodyLocKey  string   `json:"body_loc_key,omitempty"`
-	BodyLocArgs []string `json:"body_loc_args,omitempty"`
-	Icon        string   `json:"icon,omitempty"`
-	Color       string   `json:"color,omitempty"`
-	ClickAction string   `json:"click_action,omitempty"`
-	Sound       string   `json:"sound,omitempty"`
-	Image       string   `json:"image,omitempty"`
-
-	// APNS
-	Action          string   `json:"action,omitempty"`
-	ActionLocKey    string   `json:"action_loc_key,omitempty"`
-	LaunchImage     string   `json:"launch_image,omitempty"`
-	LocArgs         []string `json:"loc_args,omitempty"`
-	LocKey          string   `json:"loc_key,omitempty"`
-	Subtitle        string   `json:"subtitle,omitempty"`
-	SummaryArg      string   `json:"summary_arg,omitempty"`
-	SummaryArgCount int      `json:"summary_arg_count,omitempty"`
-}
-
-// CommonConfig is the configuration of AndroidNotification payload.
-type CommonConfig struct {
-	Enabled bool `json:"enabled,omitempty"`
-	// Common defaults for all push types.
-	commonPayload
-	// Configs for specific push types.
-	Msg commonPayload `json:"msg,omitempty"`
-	Sub commonPayload `json:"sub,omitempty"`
-}
-
-func (cp commonPayload) getStringAttr(field string) string {
-	val := reflect.ValueOf(cp).FieldByName(field)
-	if !val.IsValid() {
-		return ""
-	}
-	if val.Kind() == reflect.String {
-		return val.String()
-	}
-	return ""
-}
-
-func (cp commonPayload) getIntAttr(field string) int {
-	val := reflect.ValueOf(cp).FieldByName(field)
-	if !val.IsValid() {
-		return 0
-	}
-	if val.Kind() == reflect.Int {
-		return int(val.Int())
-	}
-	return 0
-}
-
-func (cc *CommonConfig) getStringField(what, field string) string {
-	var val string
-	if what == push.ActMsg {
-		val = cc.Msg.getStringAttr(field)
-	} else if what == push.ActSub {
-		val = cc.Sub.getStringAttr(field)
-	}
-	if val == "" {
-		val = cc.commonPayload.getStringAttr(field)
-	}
-	return val
-}
-
-func (cc *CommonConfig) getIntField(what, field string) int {
-	var val int
-	if what == push.ActMsg {
-		val = cc.Msg.getIntAttr(field)
-	} else if what == push.ActSub {
-		val = cc.Sub.getIntAttr(field)
-	}
-	if val == 0 {
-		val = cc.commonPayload.getIntAttr(field)
-	}
-	return val
-}
-
-// InterruptionLevel defines the values for the APNS payload.aps.InterruptionLevel.
-type InterruptionLevelType string
-
-const (
-	// InterruptionLevelPassive is used to indicate that notification be delivered in a passive manner.
-	InterruptionLevelPassive InterruptionLevelType = "passive"
-
-	// InterruptionLevelActive is used to indicate the importance and delivery timing of a notification.
-	InterruptionLevelActive InterruptionLevelType = "active"
-
-	// InterruptionLevelTimeSensitive is used to indicate the importance and delivery timing of a notification.
-	InterruptionLevelTimeSensitive InterruptionLevelType = "time-sensitive"
-
-	// InterruptionLevelCritical is used to indicate the importance and delivery timing of a notification.
-	// This interruption level requires an approved entitlement from Apple.
-	// See: https://developer.apple.com/documentation/usernotifications/unnotificationinterruptionlevel/
-	InterruptionLevelCritical InterruptionLevelType = "critical"
-)
-
-const (
-	// A canonical UUID that identifies the notification. If there is an error sending the notification,
-	// APNs uses this value to identify the notification to your server.
-	// The canonical form is 32 lowercase hexadecimal digits, displayed in five groups separated by hyphens
-	// in the form 8-4-4-4-12. An example UUID is as follows: 123e4567-e89b-12d3-a456-42665544000
-	// If you omit this header, a new UUID is created by APNs and returned in the response.
-	HeaderApnsID = "apns-id"
-
-	// A UNIX epoch date expressed in seconds (UTC). This header identifies the date when the notification
-	// is no longer valid and can be discarded.
-	// If this value is nonzero, APNs stores the notification and tries to deliver it at least once, repeating
-	// the attempt as needed if it is unable to deliver the notification the first time. If the value is 0,
-	// APNs treats the notification as if it expires immediately and does not store the notification or attempt
-	// to redeliver it.
-	HeaderApnsExpiration = "apns-expiration"
-
-	// The priority of the notification. Specify one of the following values:
-	// 10–Send the push message immediately. Notifications with this priority must trigger an alert, sound,
-	// or badge on the target device. It is an error to use this priority for a push notification that
-	// contains only the content-available key.
-	// 5—Send the push message at a time that takes into account power considerations for the device.
-	// Notifications with this priority might be grouped and delivered in bursts. They are throttled,
-	// and in some cases are not delivered.
-	// If you omit this header, the APNs server sets the priority to 10.
-	HeaderApnsPriority = "apns-priority"
-
-	// The topic of the remote notification, which is typically the bundle ID for your app.
-	// The certificate you create in your developer account must include the capability for this topic.
-	// If your certificate includes multiple topics, you must specify a value for this header.
-	// If you omit this request header and your APNs certificate does not specify multiple topics,
-	// the APNs server uses the certificate’s Subject as the default topic.
-	// If you are using a provider token instead of a certificate, you must specify a value for this
-	// request header. The topic you provide should be provisioned for the your team named in your developer account.
-	HeaderApnsTopic = "apns-topic"
-
-	// Multiple notifications with the same collapse identifier are displayed to the user as a single notification.
-	// The value of this key must not exceed 64 bytes. For more information, see Quality of Service,
-	// Store-and-Forward, and Coalesced Notifications.
-	HeaderApnsCollapseID = "apns-collapse-id"
-
-	// The value of this header must accurately reflect the contents of your notification’s payload.
-	// If there’s a mismatch, or if the header is missing on required systems, APNs may return an error,
-	// delay the delivery of the notification, or drop it altogether.
-	HeaderApnsPushType = "apns-push-type"
-)
-
-type ApnsPushTypeType string
-
-const (
-	// Use the alert push type for notifications that trigger a user interaction—for example, an alert, badge, or sound.
-	// If you set this push type, the apns-topic header field must use your app’s bundle ID as the topic.
-	// For more information, see Generating a remote notification.
-	// If the notification requires immediate action from the user, set notification priority to 10; otherwise use 5.
-	ApnsPushTypeAlert ApnsPushTypeType = "alert"
-
-	// Use the background push type for notifications that deliver content in the background, and don’t trigger any user interactions.
-	// If you set this push type, the apns-topic header field must use your app’s bundle ID as the topic. Always use priority 5.
-	// Using priority 10 is an error. For more information, see Pushing Background Updates to Your App.
-	ApnsPushTypeBackground ApnsPushTypeType = "background"
-
-	// Use the location push type for notifications that request a user’s location. If you set this push type,
-	// the apns-topic header field must use your app’s bundle ID with .location-query appended to the end.
-	// If the location query requires an immediate response from the Location Push Service Extension, set notification
-	// apns-priority to 10; otherwise, use 5. The location push type supports only token-based authentication.
-	ApnsPushTypeLocation ApnsPushTypeType = "location"
-
-	// Use the voip push type for notifications that provide information about an incoming Voice-over-IP (VoIP) call.
-	// For more information, see Responding to VoIP Notifications from PushKit.
-	// If you set this push type, the apns-topic header field must use your app’s bundle ID with .voip appended to the end.
-	// If you’re using certificate-based authentication, you must also register the certificate for VoIP services.
-	// The topic is then part of the 1.2.840.113635.100.6.3.4 or 1.2.840.113635.100.6.3.6 extension.
-	ApnsPushTypeVoip ApnsPushTypeType = "voip"
-
-	// Use the fileprovider push type to signal changes to a File Provider extension. If you set this push type,
-	// the apns-topic header field must use your app’s bundle ID with .pushkit.fileprovider appended to the end.
-	// For more information, see Using Push Notifications to Signal Changes.
-	ApnsPushTypeFileprovider ApnsPushTypeType = "fileprovider"
-)
-
-// aps is the APNS payload. See explanation here:
-// https://developer.apple.com/documentation/usernotifications/setting_up_a_remote_notification_server/generating_a_remote_notification#2943363
-type aps struct {
-	Alert             *apsAlert             `json:"alert,omitempty"`
-	Badge             int                   `json:"badge,omitempty"`
-	Category          string                `json:"category,omitempty"`
-	ContentAvailable  int                   `json:"content-available,omitempty"`
-	InterruptionLevel InterruptionLevelType `json:"interruption-level,omitempty"`
-	MutableContent    int                   `json:"mutable-content,omitempty"`
-	RelevanceScore    interface{}           `json:"relevance-score,omitempty"`
-	Sound             interface{}           `json:"sound,omitempty"`
-	ThreadID          string                `json:"thread-id,omitempty"`
-	URLArgs           []string              `json:"url-args,omitempty"`
-}
-
-// apsAlert is the content of the aps.Alert field.
-type apsAlert struct {
-	Action          string   `json:"action,omitempty"`
-	ActionLocKey    string   `json:"action-loc-key,omitempty"`
-	Body            string   `json:"body,omitempty"`
-	LaunchImage     string   `json:"launch-image,omitempty"`
-	LocArgs         []string `json:"loc-args,omitempty"`
-	LocKey          string   `json:"loc-key,omitempty"`
-	Title           string   `json:"title,omitempty"`
-	Subtitle        string   `json:"subtitle,omitempty"`
-	TitleLocArgs    []string `json:"title-loc-args,omitempty"`
-	TitleLocKey     string   `json:"title-loc-key,omitempty"`
-	SummaryArg      string   `json:"summary-arg,omitempty"`
-	SummaryArgCount int      `json:"summary-arg-count,omitempty"`
-}
-
-// MessageData adds user ID and device token to push message. This is needed for error handling.
-type MessageData struct {
-	Uid t.Uid
-	// FCM device token.
-	DeviceId string
-	Message  *legacy.Message
-}
 
 func payloadToData(pl *push.Payload) (map[string]string, error) {
 	if pl == nil {
@@ -454,7 +230,6 @@ func ChannelsForUser(uid t.Uid) []string {
 }
 
 func androidNotificationConfig(what, topic string, data map[string]string, config *configType) *fcmv1.AndroidConfig {
-
 	timeToLive := strconv.Itoa(defaultTimeToLive) + "s"
 	if config != nil && config.TimeToLive > 0 {
 		timeToLive = strconv.Itoa(config.TimeToLive) + "s"
@@ -487,7 +262,7 @@ func androidNotificationConfig(what, topic string, data map[string]string, confi
 		return ac
 	}
 
-	body := config.Android.getStringField(what, "Body")
+	body := config.Android.GetStringField(what, "Body")
 	if body == "$content" {
 		body = data["content"]
 	}
@@ -503,13 +278,13 @@ func androidNotificationConfig(what, topic string, data map[string]string, confi
 		Tag:                  topic,
 		NotificationPriority: priority,
 		Visibility:           "PRIVATE",
-		TitleLocKey:          config.Android.getStringField(what, "TitleLocKey"),
-		Title:                config.Android.getStringField(what, "Title"),
-		BodyLocKey:           config.Android.getStringField(what, "BodyLocKey"),
+		TitleLocKey:          config.Android.GetStringField(what, "TitleLocKey"),
+		Title:                config.Android.GetStringField(what, "Title"),
+		BodyLocKey:           config.Android.GetStringField(what, "BodyLocKey"),
 		Body:                 body,
-		Icon:                 config.Android.getStringField(what, "Icon"),
-		Color:                config.Android.getStringField(what, "Color"),
-		ClickAction:          config.Android.getStringField(what, "ClickAction"),
+		Icon:                 config.Android.GetStringField(what, "Icon"),
+		Color:                config.Android.GetStringField(what, "Color"),
+		ClickAction:          config.Android.GetStringField(what, "ClickAction"),
 	}
 
 	return ac
@@ -522,22 +297,22 @@ func apnsNotificationConfig(what, topic string, data map[string]string, unread i
 		expires = time.Now().UTC().Add(time.Duration(config.TimeToLive) * time.Second)
 	}
 	bundleId := config.ApnsBundleID
-	pushType := ApnsPushTypeAlert
+	pushType := common.ApnsPushTypeAlert
 	priority := 10
-	interruptionLevel := InterruptionLevelTimeSensitive
+	interruptionLevel := common.InterruptionLevelTimeSensitive
 	if callStatus == "started" {
 		// Send VOIP push only when a new call is started, otherwise send normal alert.
-		interruptionLevel = InterruptionLevelCritical
-		pushType = ApnsPushTypeVoip
+		interruptionLevel = common.InterruptionLevelCritical
+		pushType = common.ApnsPushTypeVoip
 		bundleId += ".voip"
 		expires = time.Now().UTC().Add(time.Duration(voipTimeToLive) * time.Second)
 	} else if what == push.ActRead {
 		priority = 5
-		interruptionLevel = InterruptionLevelPassive
-		pushType = ApnsPushTypeBackground
+		interruptionLevel = common.InterruptionLevelPassive
+		pushType = common.ApnsPushTypeBackground
 	}
 
-	apsPayload := aps{
+	apsPayload := common.Aps{
 		Badge:             unread,
 		ContentAvailable:  1,
 		MutableContent:    1,
@@ -546,22 +321,22 @@ func apnsNotificationConfig(what, topic string, data map[string]string, unread i
 	}
 
 	if config.Apns != nil && config.Apns.Enabled && what != push.ActRead {
-		body := config.Apns.getStringField(what, "Body")
+		body := config.Apns.GetStringField(what, "Body")
 		if body == "$content" {
 			body = data["content"]
 		}
 
-		apsPayload.Alert = &apsAlert{
-			Action:          config.Apns.getStringField(what, "Action"),
-			ActionLocKey:    config.Apns.getStringField(what, "ActionLocKey"),
+		apsPayload.Alert = &common.ApsAlert{
+			Action:          config.Apns.GetStringField(what, "Action"),
+			ActionLocKey:    config.Apns.GetStringField(what, "ActionLocKey"),
 			Body:            body,
-			LaunchImage:     config.Apns.getStringField(what, "LaunchImage"),
-			LocKey:          config.Apns.getStringField(what, "LocKey"),
-			Title:           config.Apns.getStringField(what, "Title"),
-			Subtitle:        config.Apns.getStringField(what, "Subtitle"),
-			TitleLocKey:     config.Apns.getStringField(what, "TitleLocKey"),
-			SummaryArg:      config.Apns.getStringField(what, "SummaryArg"),
-			SummaryArgCount: config.Apns.getIntField(what, "SummaryArgCount"),
+			LaunchImage:     config.Apns.GetStringField(what, "LaunchImage"),
+			LocKey:          config.Apns.GetStringField(what, "LocKey"),
+			Title:           config.Apns.GetStringField(what, "Title"),
+			Subtitle:        config.Apns.GetStringField(what, "Subtitle"),
+			TitleLocKey:     config.Apns.GetStringField(what, "TitleLocKey"),
+			SummaryArg:      config.Apns.GetStringField(what, "SummaryArg"),
+			SummaryArgCount: config.Apns.GetIntField(what, "SummaryArgCount"),
 		}
 	}
 
@@ -570,11 +345,11 @@ func apnsNotificationConfig(what, topic string, data map[string]string, unread i
 		return nil
 	}
 	headers := map[string]string{
-		HeaderApnsExpiration: strconv.FormatInt(expires.Unix(), 10),
-		HeaderApnsPriority:   strconv.Itoa(priority),
-		HeaderApnsTopic:      bundleId,
-		HeaderApnsCollapseID: topic,
-		HeaderApnsPushType:   string(pushType),
+		common.HeaderApnsExpiration: strconv.FormatInt(expires.Unix(), 10),
+		common.HeaderApnsPriority:   strconv.Itoa(priority),
+		common.HeaderApnsTopic:      bundleId,
+		common.HeaderApnsCollapseID: topic,
+		common.HeaderApnsPushType:   string(pushType),
 	}
 
 	ac := &fcmv1.ApnsConfig{
